Split view into separate file
